Change s3 backup target for minio
After testing and getting backups to minio working, it just makes more sense to me to put the target in this format, the bucket name can be anything a user creates in minio. Using minio as a region helps end users to have a way to know that the backup point is minio. Having something like us-west-1 can add confusion especially when the use of k8s and longhorn in on presmise.